map fx and UI HP player

+++ _Scrips/Player/Player Behaviour/HeliSlamState.cs
@@ -4,7 +4,7 @@ public class HeliSlamState : PlayerState
 {
     private float slamSpeed = 60;
     private bool hasExploded = false;
-    private int damage = 50; // Dame chiêu
+    private float damage = 50; // Dame chiêu
     private float attackRange = 1.4f; // Tầm xa chiêu
 
     public HeliSlamState(PlayerController player) : base(player) { }
